streaming/requirements.txt
kafka-python==2.0.2
feedparser==6.0.11
pyspark==4.0.1
snowflake-connector-python==3.12.2
